+++ docs/.vuepress/client.js
@@ -2,7 +2,8 @@ import { defineClientConfig } from '@vuepress/client'
 import BusTable from "./components/BusTable.vue";
 import Realtimemap from './components/Realtime-map.vue'
 import WeatherSpan from './components/weather-span.vue'
-import BusChart from './components/BusChart.vue'
+import BusChartReact from './components/BusChartReact.vue'
+import BusChartVue from './components/BusChartVue.vue'
 // import AdsenseUnit from './components/adsense-inline-article.vue'
 
 export default defineClientConfig({
@@ -10,7 +11,8 @@ export default defineClientConfig({
     app.component("BusTable", BusTable)
     app.component("Realtimemap", Realtimemap)
     app.component("WeatherSpan", WeatherSpan)
-    app.component("BusChart", BusChart)
+    app.component("BusChartReact", BusChartReact)
+    app.component("BusChartVue", BusChartVue)
     // app.component("AdsenseUnit", AdsenseUnit)
   },
 })
