+++ docs/transport/bustimer-wx.md
@@ -14,7 +14,7 @@ navbar: false
 
 <Realtimemap></Realtimemap>
 
-<BusChart/>
+<BusChartReact/>
 
 [实时线路图 by ni1o1](https://github.com/ni1o1/nikebus)
